Do not rely on ordering when stopping LRP instances

+++ handlers/desired_lrp_handlers.go
@@ -257,28 +257,30 @@ func (h *DesiredLRPHandler) stopInstancesFrom(logger lager.Logger, processGuid s
 		return
 	}
 
-	for i := index; i < len(actualLRPGroups); i++ {
+	for i := 0; i < len(actualLRPGroups); i++ {
 		group := actualLRPGroups[i]
 
 		if group.Instance != nil {
 			lrp := group.Instance
-			switch lrp.State {
-			case models.ActualLRPStateUnclaimed, models.ActualLRPStateCrashed:
-				err = h.actualLRPDB.RemoveActualLRP(logger, lrp.ProcessGuid, lrp.Index)
-				if err != nil {
-					logger.Error("failed-removing-lrp-instance", err)
-				}
-			default:
-				cellPresence, err := h.serviceClient.CellById(logger, lrp.CellId)
-				if err != nil {
-					logger.Error("failed-fetching-cell-presence", err)
-					continue
-				}
-				repClient := h.repClientFactory.CreateClient(cellPresence.RepAddress)
-				logger.Debug("stopping-lrp-instance")
-				err = repClient.StopLRPInstance(lrp.ActualLRPKey, lrp.ActualLRPInstanceKey)
-				if err != nil {
-					logger.Error("failed-stopping-lrp-instance", err)
+			if lrp.Index >= int32(index) {
+				switch lrp.State {
+				case models.ActualLRPStateUnclaimed, models.ActualLRPStateCrashed:
+					err = h.actualLRPDB.RemoveActualLRP(logger, lrp.ProcessGuid, lrp.Index)
+					if err != nil {
+						logger.Error("failed-removing-lrp-instance", err)
+					}
+				default:
+					cellPresence, err := h.serviceClient.CellById(logger, lrp.CellId)
+					if err != nil {
+						logger.Error("failed-fetching-cell-presence", err)
+						continue
+					}
+					repClient := h.repClientFactory.CreateClient(cellPresence.RepAddress)
+					logger.Debug("stopping-lrp-instance")
+					err = repClient.StopLRPInstance(lrp.ActualLRPKey, lrp.ActualLRPInstanceKey)
+					if err != nil {
+						logger.Error("failed-stopping-lrp-instance", err)
+					}
 				}
 			}
 		}
